move dir && add build pipeline
