+++ .github/workflows/docker-build.yml
@@ -13,6 +13,11 @@ on:
       - dify-agent/README.md
       - dify-agent/src/**
       - web/Dockerfile
+      - dify-agent-sandbox-runtime/docker/Dockerfile
+      - dify-agent-sandbox-runtime/go.mod
+      - dify-agent-sandbox-runtime/go.sum
+      - dify-agent-sandbox-runtime/cmd/**
+      - dify-agent-sandbox-runtime/internal/**
 
 concurrency:
   group: docker-build-${{ github.head_ref || github.run_id }}
@@ -48,6 +53,16 @@ jobs:
             runs_on: depot-ubuntu-24.04-4
             context: "{{defaultContext}}"
             file: "web/Dockerfile"
+          - service_name: "local-sandbox-amd64"
+            platform: linux/amd64
+            runs_on: depot-ubuntu-24.04-4
+            context: "{{defaultContext}}:dify-agent-sandbox-runtime"
+            file: "docker/Dockerfile"
+          - service_name: "local-sandbox-arm64"
+            platform: linux/arm64
+            runs_on: depot-ubuntu-24.04-4
+            context: "{{defaultContext}}:dify-agent-sandbox-runtime"
+            file: "docker/Dockerfile"
     steps:
       - name: Set up Depot CLI
         uses: depot/setup-action@15c09a5f77a0840ad4bce955686522a257853461 # v1.7.1
@@ -59,6 +74,7 @@ jobs:
           push: false
           context: ${{ matrix.context }}
           file: ${{ matrix.file }}
+          build-contexts: ${{ matrix.build_contexts }}
           platforms: ${{ matrix.platform }}
 
   build-docker-fork:
@@ -75,6 +91,9 @@ jobs:
           - service_name: "web-amd64"
             context: "{{defaultContext}}"
             file: "web/Dockerfile"
+          - service_name: "local-sandbox-amd64"
+            context: "{{defaultContext}}:dify-agent-sandbox-runtime"
+            file: "docker/Dockerfile"
     steps:
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@bb05f3f5519dd87d3ba754cc423b652a5edd6d2c # v4.2.0
@@ -85,4 +104,5 @@ jobs:
           push: false
           context: ${{ matrix.context }}
           file: ${{ matrix.file }}
+          build-contexts: ${{ matrix.build_contexts }}
           platforms: linux/amd64
