+++ dify-agent-sandbox-runtime/cmd/sanitize-pty/main.go
@@ -8,7 +8,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/langgenius/dify/shellctl-go/internal/sanitize"
+	"github.com/langgenius/dify/dify-agent-sandbox-runtime/internal/sanitize"
 )
 
 func main() {
