--- shellctl-go/tests/docker-compose.yml
@@ -1,41 +1,0 @@
-# docker-compose for integration testing both shellctl implementations.
-#
-# Usage:
-#   docker compose -f tests/docker-compose.yml up --build -d
-#   go test -tags=integration ./tests/... -v
-#   docker compose -f tests/docker-compose.yml down
-
-services:
-  shellctl-python:
-    build:
-      context: ../../dify-agent
-      dockerfile: docker/local-sandbox/Dockerfile
-    ports:
-      - "15004:5004"
-    environment:
-      - SHELLCTL_AUTH_TOKEN=test-token-123
-    command: ["shellctl", "serve", "--listen", "0.0.0.0:5004"]
-    healthcheck:
-      test: ["CMD", "curl", "-sf", "http://localhost:5004/healthz"]
-      interval: 2s
-      timeout: 3s
-      retries: 10
-      start_period: 10s
-
-  shellctl-go:
-    build:
-      context: ..
-      dockerfile: docker/Dockerfile
-      additional_contexts:
-        agent: ../../dify-agent
-    ports:
-      - "15005:5004"
-    environment:
-      - SHELLCTL_AUTH_TOKEN=test-token-123
-    command: ["shellctl", "serve", "--listen", "0.0.0.0:5004"]
-    healthcheck:
-      test: ["CMD", "bash", "-c", "exec 3<>/dev/tcp/localhost/5004 && echo -e 'GET /healthz HTTP/1.0\\r\\nHost: localhost\\r\\n\\r\\n' >&3 && cat <&3 | grep -q ok"]
-      interval: 2s
-      timeout: 3s
-      retries: 10
-      start_period: 5s
